refactor(catalog): move the catalog store and stats into domain catalog
catalog_store.py -> domain/catalog/store.py, endpoint_stats.py ->
domain/catalog/stats.py, with compatibility facades at the old paths.
The YAML data stays at src/treg/catalog/; the store now anchors
CATALOG_DIR at the package root instead of module-relative. A new
import-linter contract pins the catalog domain as a leaf.
