+++ src/treg/domain/catalog/__init__.py
@@ -1,0 +1,1 @@
+"""Catalog endpoints, pricing, and served-call observations."""
